Fix bug where capitalized words are missing
The file recommended in README.md that gives IPA pronunciations for
words only has the words in lowercase. Since capitalization doesn't
change pronunciation, we can check for each word in the file by checking
for the lowercase version of the word but maintaining the original
capitalization when writting the generated dictionary.