Create config.yaml to specify dictionary rules
Now for the vast majority of cases, users will only need to update a
YAML config file to customize how their dictionary is generated, instead
of having to change a Python file.

Additionally, the YAML file sets up the configuration so that it should
be straightforward to update the vowels and consonants mappings, update
phonology rules, and create many useful custom postprocessing rules.