Enable custom phoneme cluster key overrides
This adds a new feature so that in the config.yaml you can specify that
when certain phonemes appear together in a syllable, their conversion
to steno keys should be the specified keys rather than the default
behavior of combining the keys mapped to each phoneme. One way this can
be useful is to avoid out-of-steno-order errors; for example, if the
phonemes 'n' and 'tʃ' are mapped on the right-side to -PB and -FP, then
the phoneme cluster 'ntʃ' would by default be -PBF which is out of
steno order, so the generator will not be able to make translations for
words like "lunch"; but with the new overrides feature you can
specify that when right-side 'n' is followed by right-side 'tʃ' it
should get translated to a custom key cluster (e.g., -FRPB).